/W4: Fix signed/unsigned mismatch
This commit fixes warnings and bugs caused by signed/unsigned mismatch.

+++ r5dev/gameui/IConsole.cpp
@@ -384,7 +384,7 @@ void CConsole::SuggestPanel(void)
     for (size_t i = 0, ns = m_vSuggest.size(); i < ns; i++)
     {
         const CSuggest& suggest = m_vSuggest[i];
-        const bool bIsIndexActive = m_nSuggestPos == i;
+        const bool bIsIndexActive = m_nSuggestPos == ssize_t(i);
 
         ImGui::PushID(static_cast<int>(i));
